requirements.txt
# Core
python-dotenv>=1.0.1
pydantic>=2.7
pyyaml>=6.0.1

# RAG / Vector store
chromadb>=0.5.5
langchain>=0.3.7
langchain-community>=0.3.5
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.2
langgraph>=0.2.45

# Repo parsing
pathspec>=0.12.1          # .gitignore handling
gitpython>=3.1.43
pypdf>=5.0.1
python-docx>=1.1.2
markdown-it-py>=3.0.0

# Server
fastapi>=0.115.4
uvicorn[standard]>=0.32.0

# Utilities
tqdm>=4.66.5
rich>=13.9.4
